Add headers provider
